### 2020-06-29 #386 @chainsawriot
Submission: readODS
Submitting Author Name: Chung-hong Chan
Submitting Author Github Handle: @chainsawriot
Repository:    https://github.com/chainsawriot/readODS
Version submitted:  1.7.0
Version accepted: TBD

---

```
Package: readODS
Title: Read and Write ODS Files
Version: 1.7.0
Author: Gerrit-Jan Schutten, Chung-hong Chan, Thomas J. Leeper, John Foster,  and other contributors
Maintainer: Chung-hong Chan <chainsawtiney@gmail.com>
Contact: https://github.com/chainsawriot/readODS/
Description: Import ODS (OpenDocument Spreadsheet) into R as a data frame. Also support writing data frame into ODS file.
    xml2 (>= 1.3.2),
    cellranger,
    readr (>= 1.2.1),
    stringi,
    utils
    datasets,
    covr,
```

	- [ ] workflow automataion

This package reads and writes OpenDocument Spreadsheet (ods) files.

Researchers working with ods files

No
