### 2020-07-07 3005cd8
Fix problem related to appending sheets to ODS created with Libreoffice, solution by bana-ne ref #68
