+++ DESCRIPTION
@@ -21,6 +21,6 @@ Suggests:
     knitr,
     rmarkdown
 License: GPL-3
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 Roxygen: list(markdown = TRUE)
 VignetteBuilder: knitr
