### 2020-07-28 #386 @adamhsparks
- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

It's good to see the ODS format getting some attention like MS Excel's proprietary formats in R.
While ODS files may not be used as much, I still think it's important to support this file format.

I appreciate the simplicity of the package.
It does two things and it does them well.
A good UNIX philosophy.
This will help keep the package maintainable into the future and help *rio* maintenance as well, I'm sure.

There are some areas that need to be addressed before it can be accepted into rOpenSci's software community though.

Following are my comments on what I'd like to see changed and what needs to be fixed for inclusion.

###### Vignettes

* I would state that PlantGrowth dataset comes from *datasets* before illustrating how to use it to help new users that aren't familiar with R and the available resources

* For clarity, please specify that the command, `write_ods(Plant_Growth, "plant.ods")` will write the file in the current working directory of the user's session.

##### Function Documentation (and Other)

* I would suggest to include a package level help file.
`usethis::use_package_doc()` is a handy way to autogenerate the package-level documentation using details from the DESCRIPTION file.

* Please include examples for use in the helpfile documentation for `read_ods()` and `write_ods()`.

* E-mail addresses in ROxygen should use a double @ sign, e.g. `chainsawtiney@@gmail.com`.
Please refer to <http://r-pkgs.had.co.nz/man.html>, "Roxygen comments".

* Line 85 of writeODS.R, while technically correct, is unclear, especially for unfamiliar users.
I'd suggest something like "An ODS file written to the file path location specified by the user."
Or something along those lines to be clear what the return value is in plain English.

* Formatting the ROxygen comments properly will crosslink the HTML documentation, e.g. formatting "readr::type_convert" as `readr::type_convert()` on line 227 of readODS.R, will do this since you have set `Roxygen: list(markdown = TRUE)` in the DESCRIPTION file

* `write_ods()` has a parameter, `overwrite`, I think you mean to use "deprecated" in place of "depreciated" in the documentation?
This also appears on line 89 in readODS.R as "depreciated" instead of "deprecated".

* For the parameters in `write_ods()` that take a `TRUE`/`FALSE` value, I'd like to see the default value to be explicitly stated here in the documentation.

* Titles should be capitalised, "\title sections should be capitalized and not end in a period.", from <https://developer.r-project.org/Rds.html>.

##### Community Guidelines

* The DESCRIPTION file lacks a 'BugReports' field.
Most commonly this would point to https://github.com/chainsawriot/readODS/issues.

* The contributing guidelines are missing either in CONTRIBUTING or in the README.

##### General Comments on Code

* Check the output of `goodpractice::gp()` on the use of `sapply()` and `1:length()` in functions to make the code more robust.

* When checking the documentation, I relieved a warning message, `Warning message:
roxygen2 requires Encoding: UTF-8`.
Inserting `Encoding: UTF-8` in DESCRIPTION fixes the issue.

* The `path` parameter from `read_ods()` takes a default `NULL` value, but this is never checked.
If the user does not supply a sheet to read the error message simply says, `Error in .parse_ods_file(file) : object 'path' not found`.
This is honestly the most "base R" way of handling things I can think of.
But I also think it's rather confusing to new users.
This could be improved by a simple check to see if `path` is `NULL` before proceeding further when the function is executed.
A more helpful message for the user would be desirable in this case as well, *e.g.* `No file path was provided for the 'path' argument. Please provide a path to a file to import.` or something of this nature to help guide the user.

* I'm unclear on the functionality of the function on line 47 in writeODS.R.
  ```r
  .make_temp_dir <- function() {
      tmp <- tempfile()
      dir.create(tmp)
      return (tmp)
  }
  ```
Would it not be more simple to just use the `tempdir()` command here and work in that temp directory for the zipping process?
This is how I coded this functionality in the GSODR package to unzip files.

* Consider using `call. = FALSE` in `stop()` functions throughout to provide a cleaner message to end-users.

##### Installation

Installs as expected with no issues.

##### Functionality

* Imports files and writes ODS files as advertised.

* Compared with `readxl::read_excel()` it results in the exact same `data.frame()` in R.

##### Performance

* Performance is notably slower for the same data than `readxl::read_excel()`, but considering that `read_ods()` is written in R while the other is C, it makes sense and this isn't designed for speed (nor advertised for it either).

##### Automated Tests

* 93% coverage is very good, most of the missing lines aren't easily covered and/or are related to `stop()` messages.

##### Packaging Guidelines

* The assignment operators are not consistent between the functional code and the tests.
The functions use `<-` while the tests use `=`.
Per rOpenSci guidelines, either is acceptable, but you should be consistent in the use of them in the package.

* Currently the package is only checked via CI on Travis for Linux.
Using GitHub Actions and the examples from r-lib/actions, it's rather easy to set up a full check of all three major OSes and various R versions effectively.
rOpenSci guidelines require that packages are tested against the latest, previous and development versions of R.
