### 2020-07-28 ropensci/readODS#70 @chainsawriot
Issues from ropensci/software-review#386
ref ropensci/software-review#386
