### 2020-07-29 ae22503
Fixing typo of 'deprecated'
