+++ man/readODS-package.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/readODS-package.R
+\docType{package}
+\name{readODS-package}
+\alias{readODS}
+\alias{readODS-package}
+\title{readODS: Read and Write ODS Files}
+\description{
+Import ODS (OpenDocument Spreadsheet) into R as a data frame. Also support writing data frame into ODS file.
+}
+\keyword{internal}
