### 2020-07-29 chainsawriot/readODS#70 @chainsawriot
Vig

- [ ] PlantGrowth
- [ ] write_ods

Doc

- [ ] pkg-doc
- [ ] examples
- [ ] emails
- [ ] write_ods unclear
- [ ] roxy cross-link
- [ ] deprecated
- [ ] TRUE / FALSE
- [ ] Title case
