### 2020-08-12 #386 @emmamendelsohn
---

Well done--nice and simple! The code was easy to follow and review. I only have a few comments to add to what has already been mentioned.

- Consider mentioning the `list_ods_sheets()` function in the vignette/README. I sometimes use `readxl::excel_sheets()`, so I imagine the equivalent would be useful for `.ods` users.

- Perhaps not worth mentioning in the vignettes, but the `formula_as_formula` argument in `read_ods()` would have been useful when I had to QA/QC spreadsheets in my previous job :)

- In documentation for `read_ods()`, I suggest clarifying that the `na` argument can be set to `NULL`, resulting in empty cells being treated as `NA`. And perhaps rephrase "Character vector of strings to use for missing values" to "Character vector of strings to be replaced as NA".

- Check lines 92-98 in `.parse_rows()`. `read_ods("starwars.ods", skip = 11)` returns an empty data frame, when it should return the full data frame and a warning that the skip value is greater than the number of rows. The same happens when `skip = 12`. This is because in `starwars.ods`, the last two rows from `parsed_sheets` do not contain any values but are still counted toward `length(rows)`, resulting in `FALSE` for the condition on line 92.

- Not required, but you may consider giving warnings for identical and missing column names. If the user were to convert the data.frame to a tibble, these naming issues will result in errors.
