### 2020-10-05 #386 @noamross
That's fine, @chainsawriot.  We're all doing our best and are trying to be accommodating with schedules this challenging year.
