### 2020-12-27 87a9706
update Vignettes. ref #70
