### 2021-05-05 40fa055
added "logical" to all binary arguments to have more uniform documentation style
