+++ R/writeODS.R
@@ -77,10 +77,10 @@
 #' @param x a data.frame
 #' @param path Path to the ods file to write
 #' @param sheet Name of the sheet
-#' @param row_names logical, TRUE indicates that row names of x are to be included in the sheet
-#' @param col_names logical, TRUE indicates that column names of x are to be included in the sheet
 #' @param append logical, TRUE indicates that x should be appended to the existing file (path) as a new sheet. If a sheet with the same sheet_name exists, an exception is thrown. See update.
 #' @param update logical, TRUE indicates that the sheet with sheet_name in the existing file (path) should be updated with the content of x. If a sheet with sheet_name does not exist, an exception is thrown.
+#' @param row_names logical, TRUE indicates that row names of x are to be included in the sheet
+#' @param col_names logical, TRUE indicates that column names of x are to be included in the sheet
 #' @param verbose logical, if messages should be displayed
 #' @param overwrite logical, deprecated.
 #' @return the value of \code{path} invisibly.
