### 2021-05-05 344e67a
Added the default values to roxygen documentation to comply with rOpenSci
