### 2021-05-05 chainsawriot/readODS#70 @mmahmoudian
I also think the item "Title case" can be also marked as done which was supposed to be about the following reviewer's comment:

> Titles should be capitalised, "\title sections should be capitalized and not end in a period.", from https://developer.r-project.org/Rds.html.
