### 2021-05-08 8dff160
Fix for TRUE / FALSE in #70 (#73)
* added "logical" to all binary arguments to have more uniform documentation style

* Order or of arguments in documentation now match the function

* Added the default values to roxygen documentation to comply with rOpenSci
