### 2021-05-08 0bf0ea7
Generate the documentation, add mmahmoudian as a contributor
