+++ README.md
@@ -179,7 +179,8 @@ Unported License](https://creativecommons.org/licenses/by-sa/3.0/).
 The creator of this package is Gerrit-Jan Schutten. The current
 maintainer is Chung-hong Chan. This package benefits from contributions
 by Thomas J. Leeper, John Foster, Sergio Oller, Jim Hester, Stephen
-Watts, Arthur Katossky, Stas Malavin, and Duncan Garmonsway.
+Watts, Arthur Katossky, Stas Malavin, Duncan Garmonsway, and Mehrad
+Mahmoudian.
 
 This package emulates the behaviors of `readxl::read_xlsx`,
 `writexl::write_xlsx` and `xlsx::write.xlsx`.
