+++ man/read_ods.Rd
@@ -26,7 +26,7 @@ read.ods(file = NULL, sheet = NULL, formulaAsFormula = FALSE)
 
 \item{sheet}{sheet to read. Either a string (the sheet name), or an integer sheet number. The default is 1.}
 
-\item{col_names}{indicating whether the file contains the names of the variables as its first line.}
+\item{col_names}{logical, indicating whether the file contains the names of the variables as its first line. Default is TRUE.}
 
 \item{col_types}{Either NULL to guess from the spreadsheet or refer to readr::type_convert to specify cols specification. NA will return a data frame with all columns being "characters".}
 
@@ -34,15 +34,15 @@ read.ods(file = NULL, sheet = NULL, formulaAsFormula = FALSE)
 
 \item{skip}{the number of lines of the data file to skip before beginning to read data.}
 
-\item{formula_as_formula}{a switch to display formulas as formulas "SUM(A1:A3)" or as the resulting value "3"... or "8"..}
+\item{formula_as_formula}{logical, a switch to display formulas as formulas "SUM(A1:A3)" or as the resulting value "3"... or "8".. . Default is FALSE.}
 
 \item{range}{selection of rectangle using Excel-like cell range, such as \code{range = "D12:F15"} or \code{range = "R1C12:R6C15"}. Cell range processing is handled by the \code{\link[=cellranger]{cellranger}} package.}
 
-\item{row_names}{indicating whether the file contains the names of the rows as its first column}
+\item{row_names}{logical, indicating whether the file contains the names of the rows as its first column. Default is FALSE.}
 
-\item{strings_as_factors}{logical, if character columns to be converted to factors.}
+\item{strings_as_factors}{logical, if character columns to be converted to factors. Default is FALSE.}
 
-\item{verbose}{logical, if messages should be displayed.}
+\item{verbose}{logical, if messages should be displayed. Default is FALSE.}
 
 \item{file}{for read.ods only, path to the ods file.}
 
