+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: readODS
 Type: Package
 Title: Read and Write ODS Files
-Version: 1.7.1
-Date: 2020-06-22
+Version: 1.7.2
+Date: 2021-11-27
 Author: Gerrit-Jan Schutten, Chung-hong Chan, Thomas J. Leeper, John Foster,  and other contributors
 Maintainer: Chung-hong Chan <chainsawtiney@gmail.com>
 Contact: https://github.com/chainsawriot/readODS/
