### 2021-11-29 257f0ff
Clean up notes in check
