+++ NAMESPACE
@@ -7,3 +7,4 @@ export(ods_sheets)
 export(read.ods)
 export(read_ods)
 export(write_ods)
+importFrom(utils,unzip)
