### 2021-11-29 1f5ad2b
Fix all issues related to doc, ref #70
