+++ README.md
@@ -190,3 +190,23 @@ to care about file format anymore.
 ### License
 
 GPL3
+
+### Contributing
+
+Contributions in the form of feedback, comments, code, and bug report
+are welcome.
+
+  - Fork the source code, modify, and issue a [pull
+    request](https://docs.github.com/en/github/collaborating-with-issues-and-pull-requests/creating-a-pull-request-from-a-fork).
+  - Issues, bug reports: [File a Github
+    issue](https://github.com/chainsawriot/readODS).
+  - Github is not your thing? Contact Chung-hong Chan by e-mail, post,
+    or other methods listed on this
+    [page](https://www.mzes.uni-mannheim.de/d7/en/profiles/chung-hong-chan).
+
+## Code of Conduct
+
+Please note that the readODS project is released with a [Contributor
+Code of
+Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html).
+By contributing to this project, you agree to abide by its terms.
