### 2021-11-29 a6aeba9
Remove the confusing .make_temp_dir in write_ods
