### 2021-11-29 d050534
Make all `stop` with call. = FALSE
