### 2021-11-29 1392b73
Remove all : and = for assignment
