### 2021-11-29 22e2784
README: caveat about performance and file size
