+++ README.Rmd
@@ -85,7 +85,7 @@ read_ods("mtcars.ods", sheet = "plant", range = "A1:B10")
 
 ### About the speed and file size
 
-This package is written entirely in R. Although the efficiency has been improved, please don't expect the heavily optimized performance of [readxl](https://readxl.tidyverse.org/), [readr](https://readr.tidyverse.org/) and data.table's [fread](https://cran.r-project.org/web/packages/data.table/vignettes/datatable-intro.html).
+This package is written entirely in R. Although the efficiency has been improved, please don't expect the heavily optimized performance of [readxl](https://readxl.tidyverse.org/), [readr](https://readr.tidyverse.org/) and data.table's [fread](https://cran.r-project.org/package=data.table).
 
 Also, this package can't handle ODS files larger than "medium size". See [issue #71](https://github.com/chainsawriot/readODS/issues/71).
 
@@ -116,3 +116,7 @@ Contributions in the form of feedback, comments, code, and bug report are welcom
 ## Code of Conduct
 
 Please note that the readODS project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+
+```{r, include = FALSE}
+unlink("mtcars.ods")
+```
