### 2021-11-29 5b2affc
Kill most sapply (except one)
