### 2021-11-29 cfd484e
Clean the sapply with a for-loop without <<-
