+++ R/writeODS.R
@@ -49,12 +49,12 @@
 
 .find_named_sheet <- function(ss, name) {
     sheet <- NULL
-    sapply(seq(2, length(xml2::xml_children(ss))),
-           function(i) {
-               if (!is.na(xml2::xml_attr(xml2::xml_children(ss)[[i]], "name") == name) & xml2::xml_attr(xml2::xml_children(ss)[[i]], "name") == name) {
-                   sheet <<- xml2::xml_children(ss)[[i]]
-               }
-           })
+    for (i in seq(2, length(xml2::xml_children(ss)))) {
+        if (!is.na(xml2::xml_attr(xml2::xml_children(ss)[[i]], "name") == name) &
+            xml2::xml_attr(xml2::xml_children(ss)[[i]], "name") == name) {
+            sheet <- xml2::xml_children(ss)[[i]]
+        }
+    }
     return (sheet)
 }
 
