### 2022-06-07 746871e
Correct NA bug introduced by chainsawriot/readODS#77
