### 2022-06-12 b0733a3
Repeat rows bug, fixes #81, updates #77 (#82)
* Replicate repating rows

* Correct NA bug introduced by chainsawriot/readODS#77
