### 2022-06-12 a3fc836
add test cases to #82
