### 2022-06-12 8fd37d2
Revert "Repeat rows bug, fixes #81, updates #77 (#82)"
This reverts commit b0733a335859d89e99cee879e0ce355002cd2293. Due to
performance issues.
