### 2022-06-12 81d8e8f
Revert "add test cases to #82"
This reverts commit a3fc83614449f67215301f2ec9c55fa984d67eac. Also due
to performance issue.
