### 2022-06-16 e8f69d8
Add a check to prevent the checking that is not needed
Fork Matt Kerlogue's `.parse_row` to `.parse_row2` to compare
performance. Add a check to prevent nonessential checking which is
quite expensive. Can pass all previous tests.
