### 2022-06-16 d89608e
Add test for issue #81
