### 2022-06-16 3624cff
Make `.parse_row2` the default `.parse_row`
