+++ R/readODS.R
@@ -126,48 +126,6 @@
     }
     current_row <- 0
     for (row in rows) {
-        current_row <- current_row + 1
-        if (xml2::xml_has_attr(row, "table:number-rows-repeated", ods_ns)) {
-            ## empty row, just bump the current_row
-            current_row <- current_row + as.numeric(xml2::xml_attr(row, "table:number-rows-repeated", ods_ns)) - 1
-        } else {
-            ##parse the value in each column
-            current_col <- 0
-            for (cell in xml2::xml_find_all(row, ".//table:table-cell", ods_ns)) {
-                bump_cell <- .check_cell_repeat(cell, ods_ns)
-                cell_with_textp <- .check_cell_with_textp(cell, ods_ns)
-                current_col <- current_col + 1
-                if (cell_with_textp) {
-                    ## non_empty cell, get the value
-                    cell_value <- .parse_single_cell(cell, ods_ns, formula_as_formula = formula_as_formula)
-                    cell_values[[paste0(current_row, ",", current_col)]] <- cell_value
-                }
-                if (bump_cell > 1 && !cell_with_textp) {
-                    current_col <- current_col + bump_cell - 1
-                }
-                if (bump_cell > 1 && cell_with_textp) {
-                    for (bump in seq_len(bump_cell - 1)) {
-                        current_col <- current_col + 1
-                        cell_values[[paste0(current_row, ",", current_col)]] <- cell_value
-                    }
-                }
-            }
-        }
-    }
-    return(cell_values)
-}
-
-.parse_rows2 <- function(parsed_sheet, ods_ns, formula_as_formula, skip = 0) {
-    rows <- xml2::xml_find_all(parsed_sheet, ".//table:table-row", ods_ns)
-    cell_values <- new.env(hash = TRUE)
-    if (skip > 0 && skip >= length(rows)) {
-        return(cell_values)
-    }
-    if (skip > 0) {
-        rows <- rows[(skip + 1):length(rows)]
-    }
-    current_row <- 0
-    for (row in rows) {
         if (xml2::xml_has_attr(row, "table:number-rows-repeated", ods_ns)) {
             ## number of repeats
             row_repeats <- as.numeric(xml2::xml_attr(row, "table:number-rows-repeated", ods_ns))
@@ -335,7 +293,7 @@ read_ods <- function(path, sheet = 1, col_names = TRUE, col_types = NULL, na = "
     ods_ns <- res[[2]]
     sheets <- res[[1]]
     target_sheet <- .select_sheet(sheets, ods_ns = ods_ns, which_sheet = sheet)
-    cell_values <- .parse_rows2(target_sheet, ods_ns, formula_as_formula = formula_as_formula, skip = skip)
+    cell_values <- .parse_rows(target_sheet, ods_ns, formula_as_formula = formula_as_formula, skip = skip)
     parsed_df <- .convert_to_data_frame(cell_values = cell_values, header = col_names, na = na, row_header = row_names, range = range)
     ## emulate readxl to first select range.
     ## Kill unknown col_types
@@ -353,35 +311,6 @@ read_ods <- function(path, sheet = 1, col_names = TRUE, col_types = NULL, na = "
     }
     return(res)
 }
-
-read_ods2 <- function(path, sheet = 1, col_names = TRUE, col_types = NULL, na = "", skip = 0, formula_as_formula = FALSE, range = NULL,
-                     row_names = FALSE, strings_as_factors = FALSE, verbose = FALSE) {
-    if (missing(path)) {
-        stop("No file path was provided for the 'path' argument. Please provide a path to a file to import.")
-    }
[…]
