### 2022-06-16 58cc727
Fix #81 without the performance impact of #82 (#83)
* Add a check to prevent the checking that is not needed

Fork Matt Kerlogue's `.parse_row` to `.parse_row2` to compare
performance. Add a check to prevent nonessential checking which is
quite expensive. Can pass all previous tests.

* Add test for issue #81

* Make `.parse_row2` the default `.parse_row`

* Add more tests

* Make Matt Kerlogue a contributor

Co-authored-by: Matt Kerlogue <matt.kerlogue@me.com>
