+++ README.md
@@ -195,8 +195,8 @@ Unported License](https://creativecommons.org/licenses/by-sa/3.0/).
 The creator of this package is Gerrit-Jan Schutten. The current
 maintainer is Chung-hong Chan. This package benefits from contributions
 by Thomas J. Leeper, John Foster, Sergio Oller, Jim Hester, Stephen
-Watts, Arthur Katossky, Stas Malavin, Duncan Garmonsway, and Mehrad
-Mahmoudian.
+Watts, Arthur Katossky, Stas Malavin, Duncan Garmonsway, Mehrad
+Mahmoudian, and Matt Kerlogue.
 
 This package emulates the behaviors of `readxl::read_xlsx`,
 `writexl::write_xlsx` and `xlsx::write.xlsx`.
