### 2022-06-23 4dd84b2
Set rep_space to 1 if NA
