### 2022-06-24 #386 @ropensci-review-bot
Approved! Thanks @chainsawriot for submitting and @emmamendelsohn, @adamhsparks for your reviews! :grin:
